Track unsaved changes
Adds exit prompt if unsaved changes are detected. Currently ignores searchable
values inside entries, as those are not restored properly when reading the
initial entry into the model and form.

A data migration might be required as searchable values are not currently
associated with their entry in the serialized list, so they cannot be
attached properly when initializing sections.

Also adds a better back press handling mechanism, overriding (but copying)
the built-in mechanism to allow for a better API, such as allowing certain
handlers to intercept other calls and prevent back press.